# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Laravel Base — API v1",
    "version": "1.0.0",
    "description": "Kurumsal site ve admin panel base kit'inin API katmanı.\n\nWeb arayüzü ile mobil uygulamalar aynı iş mantığından beslenir: panelden değiştirilen bir menü, ayar ya da yazı iki tarafta da aynı anda değişir — API ayrı bir kopya tutmaz, aynı Service katmanını çağırır.\n\n## Yanıt biçimi\n\nHer yanıt aynı zarfı taşır. Başarıda `data`, hatada `errors` **her zaman** bulunur (içi boş olsa bile) ve boş `errors` JSON'a `{}` olarak iner, `[]` olarak değil — istemcinin aynı alanı iki ayrı tipte görmesi tipli bir modele ayrıştırmayı imkânsız kılardı.\n\n## Dil\n\nAdres dil taşımaz ve oturum yoktur. Dil şu sırayla çözülür: `?lang=` → `X-Locale` → `Accept-Language` (q değerleri ve `de-DE` gibi bölgesel etiketler dahil) → sitenin varsayılanı. Sitede olmayan bir dil **hata değildir**: varsayılana düşülür. Seçilen dil `Content-Language` ile bildirilir.\n\n## Kimlik doğrulama\n\nSanctum kişisel erişim jetonu, `Authorization: Bearer <jeton>`. Jeton **yalnızca üretildiği anda bir kez** düz metin olarak döner; veritabanında hash'li tutulur.\n\nAyrıntılı anlatım: `docs/API.md`.",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "{origin}/api/v1",
      "description": "Sürüm adreste taşınır; kırıcı bir değişiklikte /api/v2 açılır ve v1 bir süre yayında kalır.",
      "variables": {
        "origin": {
          "default": "http://localhost:8000",
...
    }
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Sanctum kişisel erişim jetonu."
      }
    },
    "parameters": {
      "AcceptLanguage": {
        "name": "Accept-Language",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "example": "tr"
        },
        "description": "Cihazın dili. q değerleri ve bölgesel etiketler (de-DE → de) dikkate alınır."
      },
      "XLocale": {
        "name": "X-Locale",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "example": "en"
        },
        "description": "Uygulama içi dil seçimi. Accept-Language'ı geçer."
      },
      "LangQuery": {
        "name": "lang",
        "in": "query",
        "required": false,
        "schema": {
          "type": "string",
          "example": "en"
        },
        "description": "En yüksek öncelikli dil seçimi."
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Jeton yok, geçersiz ya da süresi dolmuş.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "İşleme izin yok.\n\nSebebe göre `errors.code` makine tarafından okunabilir bir değer taşır: `email_unverified` (e-posta doğrulanmamış) ya da `missing_ability` (jetonun yetkisi yok, eksik yetkiler `errors.abilities` içinde). Hesap pasifse kod bulunmaz.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Doğrulama hatası.\n\n`errors` alan adından hata listesine eşlenir. `message` ilk hatayı taşır, yani istemci `errors` içine hiç bakmadan bir uyarı gösterebilir.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Hız sınırı aşıldı.\n\n`Retry-After` başlığı ne kadar beklenmesi gerektiğini söyler.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Bakım modu.\n\nKimlik uçları bakım modunda açık kalır — ön yüzde /giris de öyle.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "full_name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "avatar": {
            "$ref": "#/components/schemas/ImageSet"
          },
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "location": {
            "type": [
              "string",
              "null"
            ]
          },
          "birth_date": {
            "type": [
              "string",
              "null"
            ],
            "format": "date"
          },
          "gender": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Enum"
              },
              {
                "type": "null"
              }
            ]
          },
          "department": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Enum"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_active": {
            "type": "boolean"
          },
          "email_verified": {
            "type": "boolean"
          },
          "email_verified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            },
            "description": "Yalnız /auth/me ve profil güncelleme yanıtlarında bulunur."
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "first_name",
          "last_name",
          "full_name",
          "email",
          "is_active",
          "email_verified"
        ],
        "description": "Parola ve remember_token hiçbir koşulda dönmez."
      },
...
      "GalleryItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "type": {
            "$ref": "#/components/schemas/Enum"
          },
          "image": {
            "$ref": "#/components/schemas/ImageSet"
          },
          "video_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "duration": {
            "type": [
              "integer",
              "null"
            ]
          },
          "view_count": {
            "type": [
              "integer",
              "null"
            ]
          },
          "sort_order": {
            "type": [
              "integer",
              "null"
            ]
          },
          "locale": {
            "type": "string"
          },
          "category": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/GalleryCategory"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id"
        ]
      },
      "Slider": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          },
          "subtitle": {
            "type": [
              "string",
              "null"
            ]
          },
          "image": {
            "$ref": "#/components/schemas/ImageSet"
          },
          "button_text": {
            "type": [
              "string",
              "null"
            ]
          },
          "button_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Çözülmüş, dile duyarlı adres. Buton yoksa null."
          },
          "sort_order": {
            "type": [
              "integer",
              "null"
            ]
          },
          "locale": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
...
      "Device": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "current": {
            "type": "boolean",
            "description": "İsteği yapan oturum. Kullanıcı yanlışlıkla kendi oturumunu kapatmasın diye."
          },
          "last_used_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "current"
        ],
        "description": "Jetonun kendisi listelenmez ve listelenemez — hash'li tutulur."
      },
...
    }
  },
  "paths": {
    "/account/profile": {
      "put": {
        "tags": [
          "Hesap"
        ],
        "operationId": "updateProfile",
        "summary": "Profili güncelle",
        "description": "Güncelleme TAMDIR, parçalı değil — ad, soyad ve e-posta her istekte gönderilir. Şifre değiştirmek current_password ister. Avatar aynı istekte, çok parçalı gövdeyle gider; PHP onu yalnız POST'ta ayrıştırdığı için istemci POST + _method=PUT kullanmalı. E-POSTA DEĞİŞİRSE doğrulama sıfırlanır: yanıt başarılıdır ama bir sonraki istek 403 verir.",
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          },
          {
            "$ref": "#/components/parameters/XLocale"
          },
          {
            "$ref": "#/components/parameters/LangQuery"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "first_name": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "last_name": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 191
                  },
                  "phone": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 20
                  },
                  "current_password": {
                    "type": "string",
                    "description": "password gönderilirse zorunlu."
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "password_confirmation": {
                    "type": "string"
                  },
                  "remove_avatar": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "first_name",
                  "last_name",
                  "email"
                ]
              }
            },
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "_method": {
                    "type": "string",
                    "const": "PUT"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "description": "jpeg/png/webp, en fazla 1 MB."
                  }
                },
                "required": [
                  "_method",
                  "first_name",
                  "last_name",
                  "email"
                ]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Güncellendi.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            },
            "headers": {
              "Content-Language": {
                "schema": {
                  "type": "string"
                },
                "description": "Yanıtın hangi dilde üretildiği."
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/auth/devices": {
      "get": {
        "tags": [
          "Cihazlar"
        ],
        "operationId": "devices",
        "summary": "Açık oturumlar",
        "description": "En son kullanılan başta. Süresi dolmuş jetonlar listelenmez. Doğrulanmış e-posta istemez: hesabına şüpheli erişim olduğunu düşünen kişi oturumları kapatabilmeli.",
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          },
          {
            "$ref": "#/components/parameters/XLocale"
          },
          {
            "$ref": "#/components/parameters/LangQuery"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Oturumlar.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Device"
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "Content-Language": {
                "schema": {
                  "type": "string"
                },
                "description": "Yanıtın hangi dilde üretildiği."
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      },
      "delete": {
        "tags": [
          "Cihazlar"
        ],
        "operationId": "revokeOtherDevices",
        "summary": "Bu cihaz hariç hepsinden çık",
        "description": "Mevcut oturum KORUNUR: düğmeye basan kişi kendi uygulamasından atılmayı beklemiyor.",
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          },
          {
            "$ref": "#/components/parameters/XLocale"
          },
          {
            "$ref": "#/components/parameters/LangQuery"
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Kapatıldı.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "revoked": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "Content-Language": {
                "schema": {
                  "type": "string"
                },
                "description": "Yanıtın hangi dilde üretildiği."
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
...
    "/home": {
      "get": {
        "tags": [
          "Site"
        ],
        "operationId": "home",
        "summary": "Açılış ekranı",
        "description": "Slider + son yazılar + galeri şeridi tek istekte. Parçalar ayrı ayrı da yayında; bu uç açılıştaki üç gidiş dönüşü bire indiriyor. Bölüm başına adetler config/api.php'de.",
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          },
          {
            "$ref": "#/components/parameters/XLocale"
          },
          {
            "$ref": "#/components/parameters/LangQuery"
          }
        ],
        "responses": {
          "200": {
            "description": "Açılış verisi.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "sliders": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Slider"
                          }
                        },
                        "posts": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/BlogPost"
                          }
                        },
                        "gallery": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/GalleryItem"
                          }
                        }
                      },
                      "required": [
                        "sliders",
                        "posts",
                        "gallery"
                      ]
                    }
                  }
                }
              }
            },
            "headers": {
              "Content-Language": {
                "schema": {
                  "type": "string"
                },
                "description": "Yanıtın hangi dilde üretildiği."
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
...
    "/newsletter/subscribe": {
      "post": {
        "tags": [
          "Formlar"
        ],
        "operationId": "newsletter",
        "summary": "Bülten aboneliği",
        "description": "Varsayılan işaretli listeye düşer. Aynı adresle yeniden abone olmak yeni satır açmaz, mevcut kaydı canlandırır. Abonelikten çıkma bilerek yok: çıkış bağlantısı her kampanya mailinin altında, imzalı ve girişsiz.",
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          },
          {
            "$ref": "#/components/parameters/XLocale"
          },
          {
            "$ref": "#/components/parameters/LangQuery"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 191
                  },
                  "first_name": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 191
                  },
                  "last_name": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 191
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Abone olundu.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            },
            "headers": {
              "Content-Language": {
                "schema": {
                  "type": "string"
                },
                "description": "Yanıtın hangi dilde üretildiği."
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
...
    "/account": {
      "delete": {
        "tags": [
          "Hesap"
        ],
        "operationId": "closeAccount",
        "summary": "Hesabımı kapat",
        "description": "Hesabı yumuşak siler: kişi bir daha giriş yapamaz, bütün oturumları ve jetonları düşer, e-posta adresi serbest kalır. Şifre onayı zorunlu — jeton tek başına yetmiyor. Panele erişebilen hesaplar buradan kapanmaz (403).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/AcceptLanguage"
          },
          {
            "$ref": "#/components/parameters/XLocale"
          },
          {
            "$ref": "#/components/parameters/LangQuery"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "type": "string",
                    "maxLength": 255,
                    "description": "Mevcut şifre."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Hesap kapatıldı.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "message"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "const": true
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          }
        }
      }
    },
...
  }
}